Move database constants to constants folder
