+++ backend/infrahub/database/neo4j.py
@@ -2,9 +2,9 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
+from infrahub.constants.database import EntityType, IndexType
 from infrahub.core.query import QueryType
 
-from .constants import EntityType, IndexType
 from .index import IndexInfo, IndexItem, IndexManagerBase
 from .manager import DatabaseManager
