+++ backend/tests/query_benchmark/test_node_unique_attribute_constraint.py
@@ -3,6 +3,7 @@ from pathlib import Path
 
 import pytest
 
+from infrahub.constants.database import Neo4jRuntime
 from infrahub.core import registry
 from infrahub.core.validators.uniqueness.model import (
     NodeUniquenessQueryRequest,
@@ -11,7 +12,6 @@ from infrahub.core.validators.uniqueness.model import (
 )
 from infrahub.core.validators.uniqueness.query import NodeUniqueAttributeConstraintQuery
 from infrahub.database import QueryConfig
-from infrahub.database.constants import Neo4jRuntime
 from infrahub.log import get_logger
 from tests.helpers.constants import NEO4J_COMMUNITY_IMAGE, NEO4J_ENTERPRISE_IMAGE
 from tests.helpers.query_benchmark.benchmark_config import BenchmarkConfig
